+++ src/Library/Tinkoff/Invest/V1/Bond.php
@@ -148,13 +148,13 @@ class Bond extends \Google\Protobuf\Internal\Message
      */
     protected $aci_value = null;
     /**
-     *Код страны эмитента.
+     *Код страны риска, т.е. страны, в которой компания ведёт основной бизнес.
      *
      * Generated from protobuf field <code>string country_of_risk = 25;</code>
      */
     protected $country_of_risk = '';
     /**
-     *Наименование страны эмитента.
+     *Наименование страны риска, т.е. страны, в которой компания ведёт основной бизнес.
      *
      * Generated from protobuf field <code>string country_of_risk_name = 26;</code>
      */
@@ -289,9 +289,9 @@ class Bond extends \Google\Protobuf\Internal\Message
      *     @type \Tinkoff\Invest\V1\MoneyValue $aci_value
      *          Значение НКД (накопленного купонного дохода) на дату.
      *     @type string $country_of_risk
-     *          Код страны эмитента.
+     *          Код страны риска, т.е. страны, в которой компания ведёт основной бизнес.
      *     @type string $country_of_risk_name
-     *          Наименование страны эмитента.
+     *          Наименование страны риска, т.е. страны, в которой компания ведёт основной бизнес.
      *     @type string $sector
      *          Сектор экономики.
      *     @type string $issue_kind
@@ -898,7 +898,7 @@ class Bond extends \Google\Protobuf\Internal\Message
     }
 
     /**
-     *Код страны эмитента.
+     *Код страны риска, т.е. страны, в которой компания ведёт основной бизнес.
      *
      * Generated from protobuf field <code>string country_of_risk = 25;</code>
      * @return string
@@ -909,7 +909,7 @@ class Bond extends \Google\Protobuf\Internal\Message
     }
 
     /**
-     *Код страны эмитента.
+     *Код страны риска, т.е. страны, в которой компания ведёт основной бизнес.
      *
      * Generated from protobuf field <code>string country_of_risk = 25;</code>
      * @param string $var
@@ -924,7 +924,7 @@ class Bond extends \Google\Protobuf\Internal\Message
     }
 
     /**
-     *Наименование страны эмитента.
+     *Наименование страны риска, т.е. страны, в которой компания ведёт основной бизнес.
      *
      * Generated from protobuf field <code>string country_of_risk_name = 26;</code>
      * @return string
@@ -935,7 +935,7 @@ class Bond extends \Google\Protobuf\Internal\Message
     }
 
     /**
-     *Наименование страны эмитента.
+     *Наименование страны риска, т.е. страны, в которой компания ведёт основной бизнес.
      *
      * Generated from protobuf field <code>string country_of_risk_name = 26;</code>
      * @param string $var