+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v0.2.3 - 2022.02.19
+- [feature] Add method Get Candles
+- [feature] Add method Get Bonds
+- [feature] Add method Get Futures
+- [feature] Update api contracts
+
 ## v0.2.2 - 2022.02.13
 - [feature] Add method Get Operations
 - [feature] Add changelog